♻️refactor: 이미지 업로드 방식 변경
